requirements.txt
discord.py
discordwebhook
mysql-connector-python
pyyaml
requests
